% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamesManagement_objects.R
\name{AchievementResetMultipleForAllRequest}
\alias{AchievementResetMultipleForAllRequest}
\title{AchievementResetMultipleForAllRequest Object}
\usage{
AchievementResetMultipleForAllRequest(achievement_ids = NULL)
}
\arguments{
\item{achievement_ids}{The IDs of achievements to reset}
}
\value{
AchievementResetMultipleForAllRequest object
}
\description{
AchievementResetMultipleForAllRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for multiple achievements reset all request.
}
\seealso{
Other AchievementResetMultipleForAllRequest functions: \code{\link{achievements.resetMultipleForAllPlayers}}
}
