% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamesManagement_objects.R
\name{AchievementResetResponse}
\alias{AchievementResetResponse}
\title{AchievementResetResponse Object}
\usage{
AchievementResetResponse(currentState = NULL, definitionId = NULL,
  updateOccurred = NULL)
}
\arguments{
\item{currentState}{The current state of the achievement}

\item{definitionId}{The ID of an achievement for which player state has been updated}

\item{updateOccurred}{Flag to indicate if the requested update actually occurred}
}
\value{
AchievementResetResponse object
}
\description{
AchievementResetResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for an achievement reset response.
}

