% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamesManagement_functions.R
\name{achievements.resetAllForAllPlayers}
\alias{achievements.resetAllForAllPlayers}
\title{Resets all draft achievements for all players. This method is only available to user accounts for your developer console.}
\usage{
achievements.resetAllForAllPlayers()
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/games
\item https://www.googleapis.com/auth/plus.login
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/games, https://www.googleapis.com/auth/plus.login)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details. 

#' @importFrom googleAuthR gar_api_generator
}
\seealso{
\href{https://developers.google.com/games/services}{Google Documentation}
}
