% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamesManagement_functions.R
\name{applications.listHidden}
\alias{applications.listHidden}
\title{Get the list of players hidden from the given application. This method is only available to user accounts for your developer console.}
\usage{
applications.listHidden(applicationId, maxResults = NULL, pageToken = NULL)
}
\arguments{
\item{applicationId}{The application ID from the Google Play developer console}

\item{maxResults}{The maximum number of player resources to return in the response, used for paging}

\item{pageToken}{The token returned by the previous request}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/games
\item https://www.googleapis.com/auth/plus.login
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/games, https://www.googleapis.com/auth/plus.login)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/games/services}{Google Documentation}
}
