% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamesManagement_objects.R
\name{EventsResetMultipleForAllRequest}
\alias{EventsResetMultipleForAllRequest}
\title{EventsResetMultipleForAllRequest Object}
\usage{
EventsResetMultipleForAllRequest(event_ids = NULL)
}
\arguments{
\item{event_ids}{The IDs of events to reset}
}
\value{
EventsResetMultipleForAllRequest object
}
\description{
EventsResetMultipleForAllRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for multiple events reset all request.
}
\seealso{
Other EventsResetMultipleForAllRequest functions: \code{\link{events.resetMultipleForAllPlayers}}
}
