% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamesManagement_functions.R
\docType{package}
\name{gamesManagement_googleAuthR}
\alias{gamesManagement_googleAuthR}
\alias{gamesManagement_googleAuthR-package}
\title{Google Play Game Services Management API
The Management API for Google Play Game Services.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2016-09-03 23:14:20
filename: /Users/mark/dev/R/autoGoogleAPI/googlegamesManagementv1management.auto/R/gamesManagement_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/games
\item https://www.googleapis.com/auth/plus.login
}
}

