% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamesManagement_objects.R
\name{GamesPlayedResource}
\alias{GamesPlayedResource}
\title{GamesPlayedResource Object}
\usage{
GamesPlayedResource(autoMatched = NULL, timeMillis = NULL)
}
\arguments{
\item{autoMatched}{True if the player was auto-matched with the currently authenticated user}

\item{timeMillis}{The last time the player played the game in milliseconds since the epoch in UTC}
}
\value{
GamesPlayedResource object
}
\description{
GamesPlayedResource Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for metadata about a player playing a game with the currently authenticated user.
}

