% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamesManagement_objects.R
\name{GamesPlayerExperienceInfoResource}
\alias{GamesPlayerExperienceInfoResource}
\title{GamesPlayerExperienceInfoResource Object}
\usage{
GamesPlayerExperienceInfoResource(currentExperiencePoints = NULL,
  currentLevel = NULL, lastLevelUpTimestampMillis = NULL,
  nextLevel = NULL)
}
\arguments{
\item{currentExperiencePoints}{The current number of experience points for the player}

\item{currentLevel}{The current level of the player}

\item{lastLevelUpTimestampMillis}{The timestamp when the player was leveled up, in millis since Unix epoch UTC}

\item{nextLevel}{The next level of the player}
}
\value{
GamesPlayerExperienceInfoResource object
}
\description{
GamesPlayerExperienceInfoResource Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for 1P/3P metadata about the player's experience.
}

