% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamesManagement_objects.R
\name{HiddenPlayer}
\alias{HiddenPlayer}
\title{HiddenPlayer Object}
\usage{
HiddenPlayer(hiddenTimeMillis = NULL, player = NULL)
}
\arguments{
\item{hiddenTimeMillis}{The time this player was hidden}

\item{player}{The player information}
}
\value{
HiddenPlayer object
}
\description{
HiddenPlayer Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for the HiddenPlayer resource.
}
