% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamesManagement_objects.R
\name{HiddenPlayerList}
\alias{HiddenPlayerList}
\title{HiddenPlayerList Object}
\usage{
HiddenPlayerList(items = NULL, nextPageToken = NULL)
}
\arguments{
\item{items}{The players}

\item{nextPageToken}{The pagination token for the next page of results}
}
\value{
HiddenPlayerList object
}
\description{
HiddenPlayerList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for a list of hidden players.
}
