% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamesManagement_objects.R
\name{Player.name}
\alias{Player.name}
\title{Player.name Object}
\usage{
Player.name(familyName = NULL, givenName = NULL)
}
\arguments{
\item{familyName}{The family name of this player}

\item{givenName}{The given name of this player}
}
\value{
Player.name object
}
\description{
Player.name Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An object representation of the individual components of the player's name. For some players, these fields may not be present.
}
\seealso{
Other Player functions: \code{\link{Player}}
}
