% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamesManagement_objects.R
\name{Player}
\alias{Player}
\title{Player Object}
\usage{
Player(Player.name = NULL, avatarImageUrl = NULL,
  bannerUrlLandscape = NULL, bannerUrlPortrait = NULL, displayName = NULL,
  experienceInfo = NULL, lastPlayedWith = NULL, name = NULL,
  originalPlayerId = NULL, playerId = NULL, profileSettings = NULL,
  title = NULL)
}
\arguments{
\item{Player.name}{The \link{Player.name} object or list of objects}

\item{avatarImageUrl}{The base URL for the image that represents the player}

\item{bannerUrlLandscape}{The url to the landscape mode player banner image}

\item{bannerUrlPortrait}{The url to the portrait mode player banner image}

\item{displayName}{The name to display for the player}

\item{experienceInfo}{An object to represent Play Game experience information for the player}

\item{lastPlayedWith}{Details about the last time this player played a multiplayer game with the currently authenticated player}

\item{name}{An object representation of the individual components of the player's name}

\item{originalPlayerId}{The player ID that was used for this player the first time they signed into the game in question}

\item{playerId}{The ID of the player}

\item{profileSettings}{The player's profile settings}

\item{title}{The player's title rewarded for their game activities}
}
\value{
Player object
}
\description{
Player Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for a Player resource.
}
\seealso{
Other Player functions: \code{\link{Player.name}}
}

