% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamesManagement_objects.R
\name{PlayerScoreResetAllResponse}
\alias{PlayerScoreResetAllResponse}
\title{PlayerScoreResetAllResponse Object}
\usage{
PlayerScoreResetAllResponse(results = NULL)
}
\arguments{
\item{results}{The leaderboard reset results}
}
\value{
PlayerScoreResetAllResponse object
}
\description{
PlayerScoreResetAllResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for a list of leaderboard reset resources.
}
