% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamesManagement_objects.R
\name{PlayerScoreResetResponse}
\alias{PlayerScoreResetResponse}
\title{PlayerScoreResetResponse Object}
\usage{
PlayerScoreResetResponse(definitionId = NULL, resetScoreTimeSpans = NULL)
}
\arguments{
\item{definitionId}{The ID of an leaderboard for which player state has been updated}

\item{resetScoreTimeSpans}{The time spans of the updated score}
}
\value{
PlayerScoreResetResponse object
}
\description{
PlayerScoreResetResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for a list of reset leaderboard entry resources.
}

