% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamesManagement_functions.R
\name{quests.resetMultipleForAllPlayers}
\alias{quests.resetMultipleForAllPlayers}
\title{Resets quests with the given IDs for all players. This method is only available to user accounts for your developer console. Only draft quests may be reset.}
\usage{
quests.resetMultipleForAllPlayers(QuestsResetMultipleForAllRequest)
}
\arguments{
\item{QuestsResetMultipleForAllRequest}{The \link{QuestsResetMultipleForAllRequest} object to pass to this method
#' @importFrom googleAuthR gar_api_generator}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/games
\item https://www.googleapis.com/auth/plus.login
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/games, https://www.googleapis.com/auth/plus.login)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/games/services}{Google Documentation}

Other QuestsResetMultipleForAllRequest functions: \code{\link{QuestsResetMultipleForAllRequest}}
}

