% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamesManagement_objects.R
\name{QuestsResetMultipleForAllRequest}
\alias{QuestsResetMultipleForAllRequest}
\title{QuestsResetMultipleForAllRequest Object}
\usage{
QuestsResetMultipleForAllRequest(quest_ids = NULL)
}
\arguments{
\item{quest_ids}{The IDs of quests to reset}
}
\value{
QuestsResetMultipleForAllRequest object
}
\description{
QuestsResetMultipleForAllRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for multiple quests reset all request.
}
\seealso{
Other QuestsResetMultipleForAllRequest functions: \code{\link{quests.resetMultipleForAllPlayers}}
}
