% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamesManagement_functions.R
\name{scores.resetMultipleForAllPlayers}
\alias{scores.resetMultipleForAllPlayers}
\title{Resets scores for the leaderboards with the given IDs for all players. This method is only available to user accounts for your developer console. Only draft leaderboards may be reset.}
\usage{
scores.resetMultipleForAllPlayers(ScoresResetMultipleForAllRequest)
}
\arguments{
\item{ScoresResetMultipleForAllRequest}{The \link{ScoresResetMultipleForAllRequest} object to pass to this method
#' @importFrom googleAuthR gar_api_generator}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/games
\item https://www.googleapis.com/auth/plus.login
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/games, https://www.googleapis.com/auth/plus.login)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/games/services}{Google Documentation}

Other ScoresResetMultipleForAllRequest functions: \code{\link{ScoresResetMultipleForAllRequest}}
}
