% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamesManagement_objects.R
\name{ScoresResetMultipleForAllRequest}
\alias{ScoresResetMultipleForAllRequest}
\title{ScoresResetMultipleForAllRequest Object}
\usage{
ScoresResetMultipleForAllRequest(leaderboard_ids = NULL)
}
\arguments{
\item{leaderboard_ids}{The IDs of leaderboards to reset}
}
\value{
ScoresResetMultipleForAllRequest object
}
\description{
ScoresResetMultipleForAllRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for multiple scores reset all request.
}
\seealso{
Other ScoresResetMultipleForAllRequest functions: \code{\link{scores.resetMultipleForAllPlayers}}
}
